Update README.md
Welcome to my first GitHub project, based on a tutorial that got me started on my journey here. I seriously appreciate any feedback on it, but it's still a work-in-progress of figuring out how to do version control here, while pinning down what packaging means in programming. I was reading earlier about the branching, the community standards, and these changes relate to the latter.